o I renamed family.R to make.dmu.deta.R because this is the only thing contained in it now. The function make.dmu.deta() is now used internally but not exported.
o I changed fitfun() back to two levels of derivatives. As the dpsi 
values are only needed within biasfun(), I compute them there now.

o I changed biasfun() to not return the inverse information matrix but 
instead use it as an optional input. My feeling is that this is the more 
natural sequence of commands.

o I renamed all "ExpInfo*" objects to "Info*" in the code. When we add
supported for observed information, we then just have to make that
addition to hessfun() and we're done (and the code doesn't have any
confusingly named objects).

o I renamed the maxIter and tolerance arguments to fsmaxit and fstol
to signal that they are use for the Fisher scoring (fs).

o By setting fsmaxit = 0 one can now switch off the additional scoring.
This is now done if hessian = TRUE & type = "ML" which means that one    
really wants only the raw optim() results.

o I streamlined the Fisher scoring. In particular I pre-compute 'fit' in
each iteration and reuse it for the various components required.

o After completion of the Fisher scoring (or skipping it), I compute the
single-step BC (if requested) and then compute all quantities needed
(coef, estfun, vcov, etc.) at the optimized parameters.

o The documentation was updated to incorporate the changes above.

o Results were checked to be consistent with EJS paper on BC/BR.
 


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/betareg@111 edb9625f-4e0d-4859-8d74-9fd3b1da38cb